% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkfun.R
\name{add_logl}
\alias{add_logl}
\title{Allow users to add their own log likelihood functions}
\usage{
add_logl(funct, logl, params)
}
\arguments{
\item{funct}{objective function (must be defined in environment)}

\item{logl}{log likelihood function as expression}

\item{params}{a vector of parameters for the objective function}
}
\description{
Allow users to add their own log likelihood functions
}
