% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makevars.R
\name{make_makevars}
\alias{make_makevars}
\title{Make Makevars file to suppress irritating warnings}
\usage{
make_makevars()
}
\description{
Make Makevars file to suppress irritating warnings
}
\examples{
make_makevars()
}
