% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkfun.R
\name{mkfun}
\alias{mkfun}
\title{Deriving the log-lik and gradients}
\usage{
mkfun(formula, data)
}
\arguments{
\item{formula}{A formula in expression form of "y ~ model"}

\item{data}{A list of parameter in the formula with values in vectors}
}
\description{
Deriving the log-lik and gradients
}
