% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Deriving MLE}
\usage{
mle(form, start, data, control = mle_control())
}
\arguments{
\item{form}{A formula in expression form of "y ~ model"}

\item{start}{A list of initial values for p}

\item{data}{A list of parameter in the formula with values in vectors}

\item{control}{control arguments to pass to optimizer}
}
\description{
Deriving MLE
}
\examples{
d <- data.frame(x=0:10,y=c(26, 17, 13, 12, 20, 5, 9, 8, 5, 4, 8))
fit0 <- mle(y~dpois(lambda=ymean),start=list(ymean=mean(d$y)),data=d)
fit3 <- mle(y~dnorm(mean=ymean, sd=2),start=list(ymean=mean(d$y), ysd=2),data=d)
}
