% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle_control}
\alias{mle_control}
\title{return default values and/or user-set values for details of fitting}
\usage{
mle_control(
  optControl = list(method = "BFGS", control = list()),
  hessian_method = c("simple", "Richardson", "none")
)
}
\arguments{
\item{optControl}{list of control parameters for \code{optim}}

\item{hessian_method}{method for numerically computing Hessian}
}
\description{
return default values and/or user-set values for details of fitting
}
