% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{TMB_mkfun}
\alias{TMB_mkfun}
\title{Compiles template and create nll and gradient}
\usage{
TMB_mkfun(formula, start, links = NULL, parameters = NULL, data)
}
\arguments{
\item{formula}{A formula in expression form of "y ~ model"}

\item{start}{A list of initial values for the parameters}

\item{links}{links function and the model parameter}

\item{parameters}{??}

\item{data}{A list of parameter in the formula with values in vectors}
}
\description{
compiles template and create nll and gradient
}
