% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{TMB_template}
\alias{TMB_template}
\title{Make model template to be compiled}
\usage{
TMB_template(formula, parameter, data = NULL, link = NULL)
}
\arguments{
\item{parameter}{A list of initial values for the parameters}

\item{data}{A list of parameter in the formula with values in vectors}

\item{link}{link function and the model parameter}

\item{form}{A formula in expression form of "y ~ model"}
}
\description{
Make model template to be compiled
}
\examples{
set.seed(123)
x <- runif(20, 1, 10)
y <- rnorm(20, mean = 1.8 + 2.4 * x, sd = exp(0.3))
form <- y ~ dnorm(b0 + b1 * x, log_sigma)
links <- list(sigma="log")
TMB_template(form, parameter=list(b0=1,b1=2, log_sigma=sd(y)),
data=list(x=x,y=y),link=c(b0="identity", b1="identity", sigma="log"))
ff <- TMB_mkfun(form, parameter=list(b0=1,b1=2, log_sigma=sd(y)),
data=list(x=x,y=y),link=c(b0="identity", b1="identity", sigma="log"))
}
