% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB.R
\name{TMB_mkfun}
\alias{TMB_mkfun}
\title{Compiles template and create nll and gradient}
\usage{
TMB_mkfun(formula, parameter, data = NULL, link = NULL)
}
\arguments{
\item{parameter}{A list of initial values for the parameters}

\item{data}{A list of parameter in the formula with values in vectors}

\item{link}{link function and the model parameter}

\item{form}{A formula in expression form of "y ~ model"}
}
\description{
compiles template and create nll and gradient
}
